Avoid loading appsec mock on benchmarks
