+++ tests/bootstrap_phpbench.php
@@ -1,5 +1,7 @@
 <?php
 
+error_reporting(E_ALL);
+
 require_once __DIR__ . '/vendor/autoload.php';
 require_once __DIR__ . '/Benchmarks/vendor/autoload.php';
-require_once __DIR__ . '/bootstrap_phpunit.php';
+require_once __DIR__ . '/bootstrap_common.php';
